requirements.txt
Flask==3.1.0
google-genai>=1.51.0
gunicorn==23.0.0